Split VCom to have both Simulator and Nucleo mode. Add ARM build option.
