+++ Urchin.sln
@@ -13,40 +13,85 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "WolfCrypt", "WolfCrypt\Wolf
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|ARM = Debug|ARM
 		Debug|x64 = Debug|x64
 		Debug|x86 = Debug|x86
+		Debug-Simulator|ARM = Debug-Simulator|ARM
+		Debug-Simulator|x64 = Debug-Simulator|x64
+		Debug-Simulator|x86 = Debug-Simulator|x86
+		Release|ARM = Release|ARM
 		Release|x64 = Release|x64
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{494190CF-79B0-49D9-8DCA-BAB4CBD5A2E6}.Debug|ARM.ActiveCfg = Debug|ARM
+		{494190CF-79B0-49D9-8DCA-BAB4CBD5A2E6}.Debug|ARM.Build.0 = Debug|ARM
 		{494190CF-79B0-49D9-8DCA-BAB4CBD5A2E6}.Debug|x64.ActiveCfg = Debug|x64
 		{494190CF-79B0-49D9-8DCA-BAB4CBD5A2E6}.Debug|x64.Build.0 = Debug|x64
 		{494190CF-79B0-49D9-8DCA-BAB4CBD5A2E6}.Debug|x86.ActiveCfg = Debug|Win32
 		{494190CF-79B0-49D9-8DCA-BAB4CBD5A2E6}.Debug|x86.Build.0 = Debug|Win32
+		{494190CF-79B0-49D9-8DCA-BAB4CBD5A2E6}.Debug-Simulator|ARM.ActiveCfg = Debug-Simulator|ARM
+		{494190CF-79B0-49D9-8DCA-BAB4CBD5A2E6}.Debug-Simulator|ARM.Build.0 = Debug-Simulator|ARM
+		{494190CF-79B0-49D9-8DCA-BAB4CBD5A2E6}.Debug-Simulator|x64.ActiveCfg = Debug-Simulator|x64
+		{494190CF-79B0-49D9-8DCA-BAB4CBD5A2E6}.Debug-Simulator|x64.Build.0 = Debug-Simulator|x64
+		{494190CF-79B0-49D9-8DCA-BAB4CBD5A2E6}.Debug-Simulator|x86.ActiveCfg = Debug-Simulator|Win32
+		{494190CF-79B0-49D9-8DCA-BAB4CBD5A2E6}.Debug-Simulator|x86.Build.0 = Debug-Simulator|Win32
+		{494190CF-79B0-49D9-8DCA-BAB4CBD5A2E6}.Release|ARM.ActiveCfg = Release|ARM
+		{494190CF-79B0-49D9-8DCA-BAB4CBD5A2E6}.Release|ARM.Build.0 = Release|ARM
 		{494190CF-79B0-49D9-8DCA-BAB4CBD5A2E6}.Release|x64.ActiveCfg = Release|x64
 		{494190CF-79B0-49D9-8DCA-BAB4CBD5A2E6}.Release|x64.Build.0 = Release|x64
 		{494190CF-79B0-49D9-8DCA-BAB4CBD5A2E6}.Release|x86.ActiveCfg = Release|Win32
 		{494190CF-79B0-49D9-8DCA-BAB4CBD5A2E6}.Release|x86.Build.0 = Release|Win32
+		{B432834C-EB8F-4AA7-901F-0A7113ECC657}.Debug|ARM.ActiveCfg = Debug|ARM
+		{B432834C-EB8F-4AA7-901F-0A7113ECC657}.Debug|ARM.Build.0 = Debug|ARM
 		{B432834C-EB8F-4AA7-901F-0A7113ECC657}.Debug|x64.ActiveCfg = Debug|x64
 		{B432834C-EB8F-4AA7-901F-0A7113ECC657}.Debug|x64.Build.0 = Debug|x64
 		{B432834C-EB8F-4AA7-901F-0A7113ECC657}.Debug|x86.ActiveCfg = Debug|Win32
 		{B432834C-EB8F-4AA7-901F-0A7113ECC657}.Debug|x86.Build.0 = Debug|Win32
+		{B432834C-EB8F-4AA7-901F-0A7113ECC657}.Debug-Simulator|ARM.ActiveCfg = Debug-Simulator|ARM
+		{B432834C-EB8F-4AA7-901F-0A7113ECC657}.Debug-Simulator|ARM.Build.0 = Debug-Simulator|ARM
+		{B432834C-EB8F-4AA7-901F-0A7113ECC657}.Debug-Simulator|x64.ActiveCfg = Debug-Simulator|x64
+		{B432834C-EB8F-4AA7-901F-0A7113ECC657}.Debug-Simulator|x64.Build.0 = Debug-Simulator|x64
+		{B432834C-EB8F-4AA7-901F-0A7113ECC657}.Debug-Simulator|x86.ActiveCfg = Debug-Simulator|Win32
+		{B432834C-EB8F-4AA7-901F-0A7113ECC657}.Debug-Simulator|x86.Build.0 = Debug-Simulator|Win32
+		{B432834C-EB8F-4AA7-901F-0A7113ECC657}.Release|ARM.ActiveCfg = Release|ARM
+		{B432834C-EB8F-4AA7-901F-0A7113ECC657}.Release|ARM.Build.0 = Release|ARM
 		{B432834C-EB8F-4AA7-901F-0A7113ECC657}.Release|x64.ActiveCfg = Release|x64
 		{B432834C-EB8F-4AA7-901F-0A7113ECC657}.Release|x64.Build.0 = Release|x64
 		{B432834C-EB8F-4AA7-901F-0A7113ECC657}.Release|x86.ActiveCfg = Release|Win32
 		{B432834C-EB8F-4AA7-901F-0A7113ECC657}.Release|x86.Build.0 = Release|Win32
+		{DEDCA962-3A36-4F8A-9703-5BC0D66B961F}.Debug|ARM.ActiveCfg = Debug|ARM
+		{DEDCA962-3A36-4F8A-9703-5BC0D66B961F}.Debug|ARM.Build.0 = Debug|ARM
 		{DEDCA962-3A36-4F8A-9703-5BC0D66B961F}.Debug|x64.ActiveCfg = Debug|x64
 		{DEDCA962-3A36-4F8A-9703-5BC0D66B961F}.Debug|x64.Build.0 = Debug|x64
 		{DEDCA962-3A36-4F8A-9703-5BC0D66B961F}.Debug|x86.ActiveCfg = Debug|Win32
 		{DEDCA962-3A36-4F8A-9703-5BC0D66B961F}.Debug|x86.Build.0 = Debug|Win32
+		{DEDCA962-3A36-4F8A-9703-5BC0D66B961F}.Debug-Simulator|ARM.ActiveCfg = Debug-Simulator|ARM
+		{DEDCA962-3A36-4F8A-9703-5BC0D66B961F}.Debug-Simulator|ARM.Build.0 = Debug-Simulator|ARM
+		{DEDCA962-3A36-4F8A-9703-5BC0D66B961F}.Debug-Simulator|x64.ActiveCfg = Debug-Simulator|x64
+		{DEDCA962-3A36-4F8A-9703-5BC0D66B961F}.Debug-Simulator|x64.Build.0 = Debug-Simulator|x64
+		{DEDCA962-3A36-4F8A-9703-5BC0D66B961F}.Debug-Simulator|x86.ActiveCfg = Debug-Simulator|Win32
+		{DEDCA962-3A36-4F8A-9703-5BC0D66B961F}.Debug-Simulator|x86.Build.0 = Debug-Simulator|Win32
+		{DEDCA962-3A36-4F8A-9703-5BC0D66B961F}.Release|ARM.ActiveCfg = Release|ARM
+		{DEDCA962-3A36-4F8A-9703-5BC0D66B961F}.Release|ARM.Build.0 = Release|ARM
 		{DEDCA962-3A36-4F8A-9703-5BC0D66B961F}.Release|x64.ActiveCfg = Release|x64
 		{DEDCA962-3A36-4F8A-9703-5BC0D66B961F}.Release|x64.Build.0 = Release|x64
 		{DEDCA962-3A36-4F8A-9703-5BC0D66B961F}.Release|x86.ActiveCfg = Release|Win32
 		{DEDCA962-3A36-4F8A-9703-5BC0D66B961F}.Release|x86.Build.0 = Release|Win32
+		{6DF7134E-C549-4D03-A60B-957DA3E3FEC6}.Debug|ARM.ActiveCfg = Debug|ARM
+		{6DF7134E-C549-4D03-A60B-957DA3E3FEC6}.Debug|ARM.Build.0 = Debug|ARM
 		{6DF7134E-C549-4D03-A60B-957DA3E3FEC6}.Debug|x64.ActiveCfg = Debug|x64
 		{6DF7134E-C549-4D03-A60B-957DA3E3FEC6}.Debug|x64.Build.0 = Debug|x64
 		{6DF7134E-C549-4D03-A60B-957DA3E3FEC6}.Debug|x86.ActiveCfg = Debug|Win32
 		{6DF7134E-C549-4D03-A60B-957DA3E3FEC6}.Debug|x86.Build.0 = Debug|Win32
+		{6DF7134E-C549-4D03-A60B-957DA3E3FEC6}.Debug-Simulator|ARM.ActiveCfg = Debug|ARM
+		{6DF7134E-C549-4D03-A60B-957DA3E3FEC6}.Debug-Simulator|ARM.Build.0 = Debug|ARM
+		{6DF7134E-C549-4D03-A60B-957DA3E3FEC6}.Debug-Simulator|x64.ActiveCfg = Debug|x64
+		{6DF7134E-C549-4D03-A60B-957DA3E3FEC6}.Debug-Simulator|x64.Build.0 = Debug|x64
+		{6DF7134E-C549-4D03-A60B-957DA3E3FEC6}.Debug-Simulator|x86.ActiveCfg = Debug|Win32
+		{6DF7134E-C549-4D03-A60B-957DA3E3FEC6}.Debug-Simulator|x86.Build.0 = Debug|Win32
+		{6DF7134E-C549-4D03-A60B-957DA3E3FEC6}.Release|ARM.ActiveCfg = Release|ARM
+		{6DF7134E-C549-4D03-A60B-957DA3E3FEC6}.Release|ARM.Build.0 = Release|ARM
 		{6DF7134E-C549-4D03-A60B-957DA3E3FEC6}.Release|x64.ActiveCfg = Release|x64
 		{6DF7134E-C549-4D03-A60B-957DA3E3FEC6}.Release|x64.Build.0 = Release|x64
 		{6DF7134E-C549-4D03-A60B-957DA3E3FEC6}.Release|x86.ActiveCfg = Release|Win32
